requirements.txt
pandas
streamlit
folium
pybase64
requests
plotly_express
